requirements.txt
cssselect==1.1.0
lxml==4.6.3
parsel==1.6.0
selenium==3.141.0
six==1.15.0
speedtest-cli==2.1.3
urllib3==1.26.4
w3lib==1.22.0
